bluetooth: allow controller to be specified
If there are multiple controllers, then the wrong one might be selected on updates.
Avoid this by allowing the controller to be specified.

This also applies in the case where the bluetooth module is disabled on startup,
and we set a manual fallback path. There's no way to know the controller id beforehand,
so we default to hci0, but this might not always be the desired controller.

It would have been neat to be able to specify the controller via its MAC address,
but the way the block works now, it requires the id (hci0, hci1, etc), and as far
as I know, it's not possible to know this in all cases.